% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.mRSD.R
\name{calc.mRSD}
\alias{calc.mRSD}
\title{Title}
\usage{
calc.mRSD(
  data,
  batch = data.frame(1:nrow(data)),
  summary.range = seq(0, 100, 10),
  use = "mean"
)
}
\arguments{
\item{data}{Dataframe}

\item{batch}{Dataframe indicating batch}

\item{use}{}
}
\value{
a list
}
\description{
Title
}
