% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.dor.R
\name{calculate.dor}
\alias{calculate.dor}
\title{Calculate DoR}
\usage{
calculate.dor(xx, yy)
}
\arguments{
\item{xx}{A vector of gene1 among cells}

\item{yy}{A vector of gene2 among cells}
}
\description{
Calculate dispersion of ratio (\emph{DoR})
}
