% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidate.norm.R
\name{candidate.norm}
\alias{candidate.norm}
\title{Normalize matrix}
\usage{
candidate.norm(mat, spike_candidate, ncore = 1)
}
\arguments{
\item{mat}{Raw counts matrix, rows are genes and columns are cells}

\item{spike_candidate}{Output from \code{\link{dbscan.pick}}}

\item{ncore}{The number of cores used in parallel}
}
\value{
A list containing the normalization results for each candidate set
}
\description{
Normalize the matrix with each candidate IS gene set.
}
\examples{
\dontrun{
candidate_res<-candidate.norm(mat=mat,spike_candidate=spike_candidate,ncore=1)
}

}
