% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut.bins.R
\name{cut.bins}
\alias{cut.bins}
\title{Title: split object on bins}
\usage{
\method{cut}{bins}(obj, bins = seq(10, 100, 10))
}
\arguments{
\item{obj}{:a vector}

\item{bins}{:a sequence}
}
\value{
list of listings
}
\description{
Title: split object on bins
}
