% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbscan.pick.R
\name{dbscan.pick}
\alias{dbscan.pick}
\title{Predict IS gene sets}
\usage{
dbscan.pick(dis, ngene = (1:floor(nrow(dis)/25)) * 5, resolution = 100)
}
\arguments{
\item{dis}{The output from \code{\link{calculate.dis}}}

\item{ngene}{The number of genes expected in candidate geneset}

\item{resolution}{Parameter for scanning}
}
\value{
A list with each element containing one candidate geneset
}
\description{
Use DBscan algorithm to calculate candidate IS gene sets.
}
\examples{
\dontrun{
spike_candidate<-dbscan.pick(dis=gene_dis,ngene=(1:floor(nrow(gene_dis)/25))*5,resolution=100)
}

}
