% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.ref.R
\name{estimate.ref}
\alias{estimate.ref}
\title{Create pseudo reference cell}
\usage{
estimate.ref(mat, ref_gene)
}
\arguments{
\item{mat}{Raw counts matrix, rows are genes and columns are cells}

\item{ref_gene}{Candidate IS genes}
}
\value{
A pseudo reference cell expression
}
\description{
Create a pseudo reference cell.
}
