% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate.sf.R
\name{estimate.sf}
\alias{estimate.sf}
\title{Calculate Size factor}
\usage{
estimate.sf(cell, ref_expr)
}
\arguments{
\item{cell}{A cell expression vector}

\item{ref_expr}{A pseudo cell expression}
}
\value{
Cells size factor.
}
\description{
Calculate Size factor for cells.
}
