% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIS.R
\name{getIS}
\alias{getIS}
\title{Title}
\usage{
getIS(y, data)
}
\arguments{
\item{y}{IS name}

\item{data}{dataframe to be normalized and has IS data in it for all samples}
}
\value{
extract IS data from data.frame
}
\description{
Title
}
