% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ISnorm.R
\name{ISnorm}
\alias{ISnorm}
\title{Title : function to perform IS normalization on a data frame. It should contain IS in dataframe itself . The IS names start with word "IS"}
\usage{
ISnorm(batch_file, className, Control_samples, sample_names, Batch)
}
\arguments{
\item{batch_file}{Dataframe with compounds in rows and samples in columns}

\item{className}{list of IS names}

\item{Control_samples}{String indicating controls}

\item{sample_names}{List coming from datainfo dataframe which has all sample names}

\item{Batch}{factor indicating batch number for each sample}
}
\value{
matrix
}
\description{
Title : function to perform IS normalization on a data frame. It should contain IS in dataframe itself . The IS names start with word "IS"
}
