% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mat}
\alias{mat}
\title{Raw counts matrix}
\format{An object of class \code{data.frame} with 24175 rows and 933 columns.}
\source{
\url{https://linkinghub.elsevier.com/retrieve/pii/S0092867415005000}
}
\usage{
mat
}
\description{
A raw counts matrix with rows representing genes and columns representing cells
}
\keyword{datasets}
