% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NormalizedToIS.R
\name{NormalizedToIS}
\alias{NormalizedToIS}
\title{Function for normalizing all data with given IS}
\usage{
NormalizedToIS(class, data, listIS_vector, ALLIS)
}
\arguments{
\item{class}{:  IS name}

\item{data}{: dataframe of all data}

\item{listIS_vector}{:list for IS to be used with each lipid}

\item{ALLIS}{}
}
\value{
: IS normalized data
}
\description{
Function for normalizing all data with given IS
}
