% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt.candidate.R
\name{opt.candidate}
\alias{opt.candidate}
\title{Select best IS gene set}
\usage{
opt.candidate(mat, candidate_res, baseline_threshold = 0.1, p_value = 0.05)
}
\arguments{
\item{mat}{Raw counts matrix,rows are genes and columns are cells}

\item{candidate_res}{Output form \code{\link{candidate.norm}}}

\item{baseline_threshold}{Threshold of instability score used to choose the baseline geneset}

\item{p_value}{Threshold of \emph{p} value for \emph{F-test}}
}
\value{
Returns a list with 5 elements, respectivately normalized matrix, size factor,
 ISgenes, instability score for each cell, index of the optimized candidate geneset.
}
\description{
Automatically choose a best IS geneset based on \emph{F-test}.
}
\examples{
\dontrun{
ISnorm_res<-opt.candidate(mat=mat,candidate_res=candidate_res,baseline_threshold=0.1,p_value=0.05)
}

}
