% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.bins.R
\name{split.bins}
\alias{split.bins}
\title{Title: split object on bins}
\usage{
\method{split}{bins}(obj, bins = seq(10, 100, 10))
}
\arguments{
\item{obj}{:a vector}

\item{bins}{:a sequence}
}
\value{
list of listings
}
\description{
Title: split object on bins
}
