% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.dis.R
\name{calculate.dis}
\alias{calculate.dis}
\title{Calculate distance between genes}
\usage{
calculate.dis(mat, detection_rate = 0.9, ncore = 1, exclude = "")
}
\arguments{
\item{mat}{Raw counts matrix, rows are genes and columns are cells}

\item{detection_rate}{The threshold to filter genes, 
for example, detection_rate=0.9 means genes without at least 90\% cells having nonzero expression will not be included in further analyis.}

\item{ncore}{The number of cores used in parallel}

\item{exclude}{Genes exclusion}
}
\value{
Return a symmetrical matrix containing the distance between genes
}
\description{
Calculate pairwise distance between genes.
}
\examples{
\dontrun{
mat=read.table('raw_counts.txt')
gene_dis<-calculate.dis(mat,detection_rate=0.9,ncore=4)
}

}
